Move fixtures into tests/fixtures and split them into multiple files
