+++ app/backend/src/tests/test_admin.py
@@ -18,19 +18,16 @@ from couchers.models import (
 )
 from couchers.proto import admin_pb2, auth_pb2, events_pb2, references_pb2, reporting_pb2
 from couchers.utils import Timestamp_from_datetime, now, parse_date, timedelta
-from tests.test_communities import create_community
-from tests.test_fixtures import (
-    add_users_to_new_moderation_list,
+from tests.fixtures.db import add_users_to_new_moderation_list, generate_user, make_friends
+from tests.fixtures.misc import email_fields, mock_notification_email
+from tests.fixtures.sessions import (
     auth_api_session,
-    email_fields,
     events_session,
-    generate_user,
-    make_friends,
-    mock_notification_email,
     real_admin_session,
     references_session,
     reporting_session,
 )
+from tests.test_communities import create_community
 
 
 @pytest.fixture(autouse=True)
