+++ app/backend/src/tests/test_api.py
@@ -13,19 +13,18 @@ from couchers.proto import admin_pb2, api_pb2, blocking_pb2, jail_pb2, notificat
 from couchers.rate_limits.definitions import RATE_LIMIT_DEFINITIONS, RATE_LIMIT_HOURS
 from couchers.resources import get_badge_dict
 from couchers.utils import create_coordinate, to_aware_datetime
-from tests.test_fixtures import (
+from tests.fixtures.db import generate_user, make_friends, make_user_block
+from tests.fixtures.misc import email_fields, mock_notification_email
+from tests.fixtures.sessions import (
     api_session,
     blocking_session,
-    email_fields,
-    generate_user,
-    make_friends,
-    make_user_block,
-    mock_notification_email,
     notifications_session,
     real_api_session,
     real_jail_session,
 )
-from tests.test_fixtures import real_admin_session as admin_session
+from tests.fixtures.sessions import (
+    real_admin_session as admin_session,
+)
 
 
 @pytest.fixture(autouse=True)
