+++ app/backend/src/tests/test_auth.py
@@ -21,14 +21,9 @@ from couchers.models import (
     UserSession,
 )
 from couchers.proto import api_pb2, auth_pb2
-from tests.test_fixtures import (
-    api_session,
-    auth_api_session,
-    email_fields,
-    generate_user,
-    mock_notification_email,
-    real_api_session,
-)
+from tests.fixtures.db import generate_user
+from tests.fixtures.misc import email_fields, mock_notification_email
+from tests.fixtures.sessions import api_session, auth_api_session, real_api_session
 
 
 @pytest.fixture(autouse=True)
