+++ app/backend/src/tests/test_blocking.py
@@ -3,15 +3,12 @@ import pytest
 from google.protobuf import empty_pb2
 from sqlalchemy import select
 
+from couchers.db import session_scope
 from couchers.models import User, UserBlock
 from couchers.proto import blocking_pb2
 from couchers.servicers.blocking import is_not_visible
-from tests.test_fixtures import (
-    blocking_session,
-    generate_user,
-    make_user_block,
-    session_scope,
-)
+from tests.fixtures.db import generate_user, make_user_block
+from tests.fixtures.sessions import blocking_session
 
 
 @pytest.fixture(autouse=True)
