+++ app/backend/src/tests/test_conversations.py
@@ -19,16 +19,9 @@ from couchers.models import (
 from couchers.proto import api_pb2, conversations_pb2, notification_data_pb2, notifications_pb2
 from couchers.rate_limits.definitions import RATE_LIMIT_DEFINITIONS, RATE_LIMIT_HOURS
 from couchers.utils import Duration_from_timedelta, now, to_aware_datetime
-from tests.test_fixtures import (
-    api_session,
-    conversations_session,
-    generate_user,
-    make_friends,
-    make_user_block,
-    make_user_invisible,
-    mock_notification_email,
-    notifications_session,
-)
+from tests.fixtures.db import generate_user, make_friends, make_user_block, make_user_invisible
+from tests.fixtures.misc import mock_notification_email
+from tests.fixtures.sessions import api_session, conversations_session, notifications_session
 
 
 @pytest.fixture(autouse=True)
