+++ app/backend/src/tests/test_donations.py
@@ -12,11 +12,8 @@ from couchers.jobs.handlers import update_badges
 from couchers.models import DonationInitiation, DonationType, Invoice, InvoiceType, User, UserBadge
 from couchers.proto import donations_pb2
 from couchers.proto.google.api import httpbody_pb2
-from tests.test_fixtures import (
-    donations_session,
-    generate_user,
-    real_stripe_session,
-)
+from tests.fixtures.db import generate_user
+from tests.fixtures.sessions import donations_session, real_stripe_session
 
 
 @pytest.fixture(autouse=True)
