+++ app/backend/src/tests/test_editor.py
@@ -11,10 +11,8 @@ from couchers.models import (
     Volunteer,
 )
 from couchers.proto import editor_pb2
-from tests.test_fixtures import (
-    generate_user,
-    real_editor_session,
-)
+from tests.fixtures.db import generate_user
+from tests.fixtures.sessions import real_editor_session
 
 
 @pytest.fixture(autouse=True)
