+++ app/backend/src/tests/test_email.py
@@ -28,18 +28,10 @@ from couchers.tasks import (
     send_signup_email,
 )
 from couchers.utils import Timestamp_from_datetime, now, timedelta
+from tests.fixtures.db import generate_user
+from tests.fixtures.misc import email_fields, mock_notification_email, process_jobs
+from tests.fixtures.sessions import api_session, events_session, notifications_session, real_editor_session
 from tests.test_communities import create_community
-from tests.test_fixtures import (
-    api_session,
-    email_fields,
-    events_session,
-    generate_user,
-    mock_notification_email,
-    notifications_session,
-    process_jobs,
-    real_editor_session,
-    session_scope,
-)
 
 
 @pytest.fixture(autouse=True)
