+++ app/backend/src/tests/test_galleries.py
@@ -6,10 +6,8 @@ from sqlalchemy.exc import IntegrityError
 from couchers.db import session_scope
 from couchers.models import PhotoGallery, PhotoGalleryItem, Upload, User
 from couchers.proto import galleries_pb2
-from tests.test_fixtures import (
-    galleries_session,
-    generate_user,
-)
+from tests.fixtures.db import generate_user
+from tests.fixtures.sessions import galleries_session
 
 
 @pytest.fixture(autouse=True)
