+++ app/backend/src/tests/test_gis.py
@@ -4,11 +4,9 @@ import pytest
 from google.protobuf import empty_pb2
 
 from couchers.materialized_views import refresh_materialized_views, refresh_materialized_views_rapid
+from tests.fixtures.db import generate_user
+from tests.fixtures.sessions import gis_session
 from tests.test_communities import testing_communities  # noqa
-from tests.test_fixtures import (
-    generate_user,
-    gis_session,
-)
 
 
 @pytest.fixture(autouse=True)
