+++ app/backend/src/tests/test_interceptors.py
@@ -17,10 +17,8 @@ from couchers.models import APICall, UserSession
 from couchers.proto import account_pb2, admin_pb2, api_pb2, auth_pb2
 from couchers.servicers.account import Account
 from couchers.servicers.api import API
-from tests.test_fixtures import (
-    generate_user,
-    real_admin_session,
-)
+from tests.fixtures.db import generate_user
+from tests.fixtures.sessions import real_admin_session
 
 
 @pytest.fixture(autouse=True)
