+++ app/backend/src/tests/test_media.py
@@ -8,11 +8,8 @@ from couchers.crypto import b64decode, random_hex
 from couchers.db import session_scope
 from couchers.models import InitiatedUpload, Upload
 from couchers.proto import media_pb2
-from tests.test_fixtures import (
-    api_session,
-    generate_user,
-    media_session,
-)
+from tests.fixtures.db import generate_user
+from tests.fixtures.sessions import api_session, media_session
 
 
 @pytest.fixture(autouse=True)
