+++ app/backend/src/tests/test_search.py
@@ -8,13 +8,9 @@ from couchers.materialized_views import refresh_materialized_views, refresh_mate
 from couchers.models import EventOccurrence, HostingStatus, LanguageAbility, LanguageFluency, MeetupStatus
 from couchers.proto import api_pb2, communities_pb2, events_pb2, search_pb2
 from couchers.utils import Timestamp_from_datetime, create_coordinate, millis_from_dt, now
+from tests.fixtures.db import generate_user
+from tests.fixtures.sessions import communities_session, events_session, search_session
 from tests.test_communities import create_community, testing_communities  # noqa
-from tests.test_fixtures import (
-    communities_session,
-    events_session,
-    generate_user,
-    search_session,
-)
 from tests.test_references import create_friend_reference
